Type check for delete?

+++ admin_page.php
@@ -41,7 +41,7 @@ This is the login screen
                         <form action="admin_page.php" method="POST" autocomplete="off">
                         	Delete Player
                             <div class="form-group">
-                                <input type="text" class="form-control" name="player_id" width="10">
+                                <input type="text" class="form-control" name="player_id" placeholder="Enter Player ID" width="10">
                             </div>
                             <input type="submit" id="deletePlayer" value="Delete" name="delete" class="btn btn-primary">
                         </form>
@@ -52,15 +52,27 @@ This is the login screen
 		$db_conn = oci_connect("ora_y0w0b", "a21529145", "dbhost.ugrad.cs.ubc.ca:1522/ug");
 		if($db_conn){
 
+		if($_SESSION["ERROR"]){
+			$_SESSION["ERROR"] = 0;
+			echo "<script type='text/javascript'>alert('User Does Not Exist!!!');</script>";			
+		}
+
 		echo "<script>console.log( 'DB Connected' );</script>";
 		if(array_key_exists("delete", $_POST)){
 			$player_id = $_POST['player_id'];
+			if(is_int($player_id)){
 			$_SESSION["player_id"] = $player_id;
 			$query = "delete from Player where Player_ID = $player_id";
 			$result = executePlainSQL($query);
 			OCICommit($db_conn);
 			oci_free_statement($result);
+			$_SESSION["ERROR"] = 0;
+			}
+			else{
+				$_SESSION["ERROR"] = 1;
+			}
 			header('Location: http://www.ugrad.cs.ubc.ca/~s4i0b/SquishyBoots/admin_page.php');
+
 		}
 		//Show the Character Table
 		else {
